requirements.txt
python-telegram-bot==20.3
requests==2.31.0
pyopenssl==23.2.0
cryptography==41.0.2
python-dotenv==1.0.0
python-dateutil==2.8.2